refactor: extract the processDiagram method
